refactor: extract whole methods from Application Class and move them to new 3 classes
하나의 클래스에 모든 함수들이 포함되어 있어 리팩토링을 위해 Game, Functions, UserInput 라는 3개의 클래스를 만들었다.
Game class는 프로그램의 전체적인 진행과 관련된 method들로 구성되어 있다.
Functions class는 게임에서 필요한 기능들을 구현한 method들로 구성되어 있다.
UserInput class는 유저가 데이터를 입력할 때 필요한 기능들을 구현한 method들로 구성되어 있다.
